data_pipeline/requirements.txt
duckdb>=0.10.0
boto3>=1.34.0
pyarrow>=15.0.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
